requirements.txt
setuptools>=65.0.0
Flask==3.0.0
gunicorn==21.2.0
Jinja2==3.1.2
itsdangerous==2.2.0
click==8.3.3
blinker==1.9.0
charset-normalizer==3.4.7
reportlab
qrcode[pil]
psycopg2-binary